% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{confusion_mat}
\alias{confusion_mat}
\title{Compute metrics and compute confusion matrix}
\usage{
confusion_mat(pred, true = NULL, title, file.out = F)
}
\arguments{
\item{pred}{A vector of factor 0,1 as the predicted labels}

\item{true}{A vector of factor 0,1 as the predicted labels. If NULL, use y_test}

\item{title}{a character specifying the plot title}

\item{file.out}{NULL if not saving plot locally. A character specifing the location
to save the output plot.}
}
\value{
A ggplot object.
}
\description{
This function returns the output of caret's confusionMatrix and, if specified, plots the
confusion matrix heatmap
}
