% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{phishing_corrs_plot}
\alias{phishing_corrs_plot}
\title{Phishing Data Correlation plot}
\usage{
phishing_corrs_plot()
}
\value{
A ggplot object of the correlation plot
}
\description{
This function returns a correlation plot based on all variables in the original phishing data
}
\examples{

phishing_corrs_plot()

}
