% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_data_checks.R
\name{phishing_data_checks}
\alias{phishing_data_checks}
\title{Basic checks of the original source data set.}
\usage{
phishing_data_checks()
}
\description{
This function checks for basic characteristics about the original data set as part of EDA.
}
\examples{

phishing_data_checks()

}
