% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{phishing_pca}
\alias{phishing_pca}
\title{Conduct PCA on original full phishing data}
\usage{
phishing_pca()
}
\value{
A ggplot object of a plot of PC1 vs PC2 augmented by true class labels
}
\description{
This function returns a plot of PC1 vs PC2 augmented by true class labels.
}
\examples{

phishing_pca()

}
