% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{plot_rf_cv}
\alias{plot_rf_cv}
\title{Plot parameter tuning results}
\usage{
plot_rf_cv(file.in, title, file.out = NULL)
}
\arguments{
\item{file.in}{a dataframe of the CV result}

\item{title}{a character specifying the plot title}

\item{file.out}{NULL if not saving plot locally. A character specifing the location
to save the output plot.}
}
\value{
A ggplot object.
}
\description{
This function returns and (if specified) saves the scatter plot of Kappa
for different parameter combinations of random forest model fitted on phishing dataset's
training set after k-fold CV.
}
\examples{

plot_rf_cv(rf_full_cv,"test plot")

}
