% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_full_cv.R
\docType{data}
\name{rf_full_cv}
\alias{rf_full_cv}
\title{RF Full model CV result}
\format{
A data frame with parameter combinations and CV metrics
}
\usage{
rf_full_cv
}
\description{
The cross validation result of RF full model
}
\keyword{datasets}
