% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_reduce_cv.R
\docType{data}
\name{rf_reduce_cv}
\alias{rf_reduce_cv}
\title{RF Reduced model CV result}
\format{
A data frame with parameter combinations and CV metrics
}
\usage{
rf_reduce_cv
}
\description{
The cross validation result of RF reduced model
}
\keyword{datasets}
