% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{run_rf}
\alias{run_rf}
\title{Run random forest, output the optimal model, and save parameter tuning result.}
\usage{
run_rf(
  k = 5,
  cv = T,
  mtry = c(2, 10, 19),
  min.node.size = c(1, 5),
  n_tree = c(5, 50, 500),
  splitrule = c("gini", "extratrees"),
  seed = 123,
  verbose = F,
  trimmed_var = NULL,
  file = NULL
)
}
\arguments{
\item{k}{an integer for CV's fold number}

\item{cv}{an boolean indicating if running CV for parameter selection}

\item{mtry}{When cv=T: a vector of integers; when cv=F: a integer. Specifies choice of mtry for RF.}

\item{min.node.size}{When cv=T: a vector of integers; when cv=F: a integer. Specifies choice of mim.node.size for RF.}

\item{n_tree}{When cv=T: a vector of integers; when cv=F: a integer. Specifies choice of n_tree for RF.}

\item{seed}{An integer specifying the random seed for model fitting.}

\item{verbose}{a boolean indicating if verbose for CV iterations should be printed out.}

\item{trimmed_var}{NULL for none of a vector of strings indicating features to
be removed from x when training model.}

\item{file}{NULL for no file output, or a string indicating the position where
the CV result is saved at}
}
\value{
A "train" object from caret: the optimal model.
}
\description{
This function runs random forest classification on the training set of
phishing data. k-fold CV will be run through the user-provided parameter
grid. A dataframe of CV will be saved if a local dir is specified. Ouput is the model
with the highest Kappa in CV, then fitted based on the complete training set data.
}
\examples{

run_rf(file='/path/to/output')

}
