% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{run_svm}
\alias{run_svm}
\title{Using best performed tuning parameter traning svm model and output performance on test set}
\usage{
run_svm(
  seed = seed,
  reduced,
  RBF,
  train_x = train_x,
  train_y = train_y,
  test_x = test_x,
  test_y = test_y
)
}
\arguments{
\item{seed}{An integer specifying the random seed for model fitting.}

\item{reduced}{logical,whether run reduced model, which remove unimportant variables}

\item{RBF}{logical, whether fit RBF svm model. If False, fit linear RBF model.}

\item{train_x}{matrix, design matrix in training data}

\item{train_y}{vector, outcome in training data}

\item{test_x}{matrix, design matrix in testing data}

\item{test_y}{vector, outcome in testing data}
}
\value{
a list of index of model performance in test set: Sensitivity,Specificity,f1_score,kappa
}
\description{
This function uses pre-tunned hyperparameters for Linear or RBF svm model to output Sensitivity,Specificity,f1_score,kappa of the model in the test set.
}
\examples{

run_svm(111,reduced = T,RBF=F,train_x=train_x,train_y=train_y,test_x=test_x,test_y=test_y)

}
