% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGD_logistic.R
\name{SGD_logistic}
\alias{SGD_logistic}
\title{Mini-Batch Logistic Regression}
\usage{
SGD_logistic(
  y,
  X,
  batch_size = 50,
  learning_rate = 0.001,
  epsilon = 1e-06,
  max_iter = 20000
)
}
\arguments{
\item{y}{Numeric vector of binary response variables.}

\item{X}{Numeric matrix of predictor variables.}

\item{batch_size}{Size of each batch.}

\item{learning_rate}{Learning rate for gradient descent.}

\item{epsilon}{Convergence criterion for stopping gradient descent.}

\item{max_iter}{Maximum number of iterations.}
}
\value{
A list containing the coefficients and the number of iterations.
}
\description{
Performs logistic regression using mini-batch gradient descent.
}
