% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\docType{data}
\name{source_data}
\alias{source_data}
\title{Web phishing data set}
\format{
A data frame with 20 columns
}
\source{
https://www.kaggle.com/datasets/danielfernandon/web-page-phishing-dataset
}
\usage{
source_data
}
\description{
The original Web phishing data for Bios735 Group 1's final project
}
\keyword{datasets}
