% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitting.R
\name{split_phishing_data}
\alias{split_phishing_data}
\title{Split data into train and test sets}
\usage{
split_phishing_data()
}
\value{
train_x, train_y, test_x, test_y csv files written to derived_data directory.
}
\description{
This function splits original phishing data into train and test sets and writes them to csv files.
}
\examples{

split_phishing_data()

}
